// src/app/models/models.ts
export type Product = {
  id: number;
  name: string;
  price: number;
  url: string;
  description: string;
};
export type CProduct = {
  id: number;
  name: string;
  price: number;
  url: string;
  description: string;
  amount: number;
};
export type Ouser = {
  name: string;
  address: string;
  ccn: string;
};